requirements.txt
django
djangorestframework
django-cors-headers
celery
rabbitmq
drf-yasg
django-cors-headers 
django-environ
mysqlclient
django-chapa